Add files via upload
I made an original method of finding starting price and ending price by measuring the date. I haven't tested it yet but I wanted to save it first. The original directions were to find the starting price by seeing if the previous line is the same or not. As well as seeing if the following line is the same. I based my code on a spreadsheet that is not in order. So I based my starting price on the date instead.